refactor(RankingCounter): 메인 로직 서비스에서 도메인으로 이동

+++ src/main/java/lotto/dto/StatisticsResult.java
@@ -2,5 +2,5 @@ package lotto.dto;
 
 import java.util.Map;
 
-public record StatisticsResult(Map<String, Integer> prizeCount, int prizeMoney, int payment) {
+public record StatisticsResult(Map<String, Integer> prizeCount, long prizeMoney, int payment) {
 }
